+++ pythello/utils/precondition.py
